fix: review comments impl
